Update rest_api.rst
It seems that in FW 2.5.9 there's an additional parameter in /xled/v1/led/mode: shop_mode.
Result of GET is {"mode":"off","shop_mode":0,"code":1000}

When shop_mode is set to 1, the LED on the controller with the HW button is continuously changing in this sequence: red, yellow(ish), green, blue with approximately 0.5 sec delay. You can set any mode together with shop_mode 1, but the mode is automatically set to "demo" about 15 seconds after the command is issued. Within the 15 minute the HW button can be used to switch the unit into "effect" mode, but this is also automatically overridden with "demo" mode after 15 sec passed.
shop_mode can be set to any number between 0 and 255, I didn't notice any difference. When higher or lower  value is entered, a modulo 256 operation converts it to a number between 0 and 255.

With the app you can still control the lights. Until an effect is not applied it changes temporarily without exiting shop_mode. As soon as you apply the effect, shop_mode obviously set back to 0.

You are allowed to send only the mode parameter without shop_mode defined.